+++ src/reducers/settingsReducer.ts
@@ -12,12 +12,14 @@ export interface SettingsState {
     forceMcuBoot: boolean;
     autoRead: boolean;
     autoReset: boolean;
+    autoUpdateOBFirmware: boolean;
 }
 
 const initialState: SettingsState = {
     forceMcuBoot: false,
     autoRead: false,
     autoReset: false,
+    autoUpdateOBFirmware: false,
 };
 
 const settingsSlice = createSlice({
@@ -27,6 +29,7 @@ const settingsSlice = createSlice({
         settingsLoad(state, action: PayloadAction<SettingsState>) {
             state.autoRead = action.payload.autoRead;
             state.autoReset = action.payload.autoReset;
+            state.autoUpdateOBFirmware = action.payload.autoUpdateOBFirmware;
         },
         toggleAutoRead(state) {
             state.autoRead = !state.autoRead;
@@ -34,14 +37,29 @@ const settingsSlice = createSlice({
         toggleAutoReset(state) {
             state.autoReset = !state.autoReset;
         },
+        toggleAutoUpdateOBFirmware(state) {
+            state.autoUpdateOBFirmware = !state.autoUpdateOBFirmware;
+        },
     },
 });
 
 export default settingsSlice.reducer;
 
-const { settingsLoad, toggleAutoRead, toggleAutoReset } = settingsSlice.actions;
+const {
+    settingsLoad,
+    toggleAutoRead,
+    toggleAutoReset,
+    toggleAutoUpdateOBFirmware,
+} = settingsSlice.actions;
 
-export { settingsLoad, toggleAutoRead, toggleAutoReset };
+export {
+    settingsLoad,
+    toggleAutoRead,
+    toggleAutoReset,
+    toggleAutoUpdateOBFirmware,
+};
 
 export const getAutoRead = (state: RootState) => state.app.settings.autoRead;
 export const getAutoReset = (state: RootState) => state.app.settings.autoReset;
+export const getAutoUpdateOBFirmware = (state: RootState) =>
+    state.app.settings.autoUpdateOBFirmware;