+++ Changelog.md
@@ -1,4 +1,10 @@
-## 4.3.1 - Unreleased
+## 4.4.0 Unreleased
+
+### Added
+
+-   J-Link devices OB FW can now be automatically updated when erasing, writing,
+    and reading.
+-   Button to manually update the OB Firmware of a device.
 
 ### Changed